feat : implemented createEvent, changed reissueToken, moved CustomExeptionHandler code to GlobalExceptionHandler
