+++ src/main/java/com/sosim/server/jwt/JwtController.java
@@ -2,15 +2,14 @@ package com.sosim.server.jwt;
 
 import com.sosim.server.common.response.Response;
 import com.sosim.server.jwt.dto.ReIssueTokenInfo;
-import com.sosim.server.jwt.dto.ReIssueTokenReq;
 import com.sosim.server.type.CodeType;
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -20,10 +19,10 @@ public class JwtController {
 
     private final JwtService jwtService;
 
-    @PostMapping("/login/reissueToken")
-    public ResponseEntity<?> reIssueToken(@RequestBody ReIssueTokenReq reIssueTokenReq, HttpServletResponse httpServletResponse) {
+    @GetMapping("/login/reissueToken")
+    public ResponseEntity<?> reIssueToken(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse) {
 
-        ReIssueTokenInfo reIssueTokenInfo = this.jwtService.verifyRefreshTokenAndReIssueAccessToken(reIssueTokenReq, httpServletResponse);
+        ReIssueTokenInfo reIssueTokenInfo = this.jwtService.verifyRefreshTokenAndReIssueAccessToken(httpServletRequest, httpServletResponse);
         if (reIssueTokenInfo == null) {
             return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
         }
