+++ src/main/java/com/sosim/server/jwt/JwtService.java
@@ -1,7 +1,6 @@
 package com.sosim.server.jwt;
 
 import com.sosim.server.jwt.dto.ReIssueTokenInfo;
-import com.sosim.server.jwt.dto.ReIssueTokenReq;
 import com.sosim.server.user.User;
 import java.io.IOException;
 import javax.servlet.FilterChain;
@@ -12,7 +11,7 @@ import javax.servlet.http.HttpServletResponse;
 public interface JwtService {
 
     void saveRefreshToken(RefreshToken refreshToken);
-    ReIssueTokenInfo verifyRefreshTokenAndReIssueAccessToken(ReIssueTokenReq reIssueTokenReq, HttpServletResponse response);
+    ReIssueTokenInfo verifyRefreshTokenAndReIssueAccessToken(HttpServletRequest httpServletRequest, HttpServletResponse response);
     void sendRefreshToken(HttpServletResponse response, String refreshToken);
     void checkAccessTokenAndAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
         throws ServletException, IOException;
